BREAKING: A new strategy to handle the observable being re-subscribed to after all subscribers have previously unsubscribed.
With the previous strategy, if a live stream was unsubscribed from
and then re-subscribed to, the subscriber would not have been able to
see the values it has missed inbetween.

The new approach allows the caller to specify how to resume the stream
from the appropriate point by taking a function that accepts the latest
value and return a new stream. See the README.md document for more
details.